decouple the code to get the aggregate result from state
